Change 'X' to varchar(4000) as per your own docs
This http://www.concrete5.org/documentation/how-tos/developers/creating-and-working-with-db-xml-files/ says that 'X' is a varchar capped at length 4000, and my application relies on that behavior. I therefore changed X from 'text' to 'string' and additionally introduced an explicit length limitation for the 'X' and 'C' types so it complies with the documentation on the above webpage.